test: LottoTest 로또 당첨 번호와 로또 번호 일치 개수 테스트 추가 - 로또 번호와 당첨 번호의 일치 개수 계산을 테스트 한다.
